refactor(web): extract comment-composer from task detail
